implement paging.
TODO: Need to fix issues and check performance

+++ features/home/home-data/src/main/java/com/bestway/data/repositories/homerepository/HomeRepositoryImpl.kt
@@ -13,20 +13,20 @@ class HomeRepositoryImpl(
     private val redditPostDao: RedditPostDao,
     private val homeService: HomeService
 ) : HomeRepository {
-    override fun getHotPosts(shouldRefreshData: Boolean): Flow<List<RedditPost>?> {
+    override fun getHotPosts(shouldRefreshData: Boolean, nextPageKey: String?): Flow<List<RedditPost>?> {
         return flow {
             val allPostsFromDb = redditPostDao.getAllRedditPosts()
 
             emit(
-                if (allPostsFromDb.isNotEmpty() && !shouldRefreshData) {
+                if (allPostsFromDb.isNotEmpty() && !shouldRefreshData && nextPageKey.isNullOrEmpty()) {
                     // If there are posts in the database and no refresh is needed, emit them directly.
                     allPostsFromDb.asDomain()
                 } else {
                     // If a refresh is needed or there are no posts in the database, fetch new data.
-                    val listings = homeService.getHotListings().getOrThrow().asEntity()
+                    val listings = homeService.getHotListings(nextPageKey = nextPageKey).getOrThrow().asEntity()
 
-                    if (allPostsFromDb.isNotEmpty()) {
-                        // If there are existing posts in the database, delete them before inserting new ones.
+                    if (allPostsFromDb.isNotEmpty() && shouldRefreshData) {
+                        // If there are existing posts in the database and refresh is needed, delete them before inserting new ones.
                         redditPostDao.deleteAllRedditPosts()
                     }
